+++ src/Modele/Joueurs/JoueurPhysique.java
@@ -1,4 +1,4 @@
-package Joueurs;
+package Modele.Joueurs;
 
 import Modele.Plateau.Pingouin;
 import Modele.Plateau.Plateau;
